+++ lib/processor/constructor.go
@@ -119,6 +119,7 @@ const (
 	TypeProcessMap   = "process_map"
 	TypeRateLimit    = "rate_limit"
 	TypeRedis        = "redis"
+	TypeResource     = "resource"
 	TypeSample       = "sample"
 	TypeSelectParts  = "select_parts"
 	TypeSleep        = "sleep"
@@ -180,6 +181,7 @@ type Config struct {
 	ProcessMap   ProcessMapConfig   `json:"process_map" yaml:"process_map"`
 	RateLimit    RateLimitConfig    `json:"rate_limit" yaml:"rate_limit"`
 	Redis        RedisConfig        `json:"redis" yaml:"redis"`
+	Resource     string             `json:"resource" yaml:"resource"`
 	Sample       SampleConfig       `json:"sample" yaml:"sample"`
 	SelectParts  SelectPartsConfig  `json:"select_parts" yaml:"select_parts"`
 	Sleep        SleepConfig        `json:"sleep" yaml:"sleep"`
@@ -240,6 +242,7 @@ func NewConfig() Config {
 		ProcessMap:   NewProcessMapConfig(),
 		RateLimit:    NewRateLimitConfig(),
 		Redis:        NewRedisConfig(),
+		Resource:     "",
 		Sample:       NewSampleConfig(),
 		SelectParts:  NewSelectPartsConfig(),
 		Sleep:        NewSleepConfig(),